+++ packages/components/src/Drawer/stories/SampleContent.tsx
@@ -26,10 +26,10 @@
 
 import React, { useContext } from 'react'
 import styled from 'styled-components'
-import { DialogHeader, DialogContent, DialogFooter } from '../../Layout'
-import { DialogContext } from '../../DialogContext'
-import { Button, ButtonTransparent } from '../../../Button'
-import { Heading, Paragraph } from '../../../Text'
+import { DialogHeader, DialogContent, DialogFooter } from '../../Dialog/Layout'
+import { DialogContext } from '../../Dialog/DialogContext'
+import { Button, ButtonTransparent } from '../../Button'
+import { Heading, Paragraph } from '../../Text'
 
 export const SampleContent = () => {
   const { closeModal } = useContext(DialogContext)
