+++ packages/components/src/index.ts
@@ -32,6 +32,7 @@ export * from './Button'
 export * from './Card'
 export * from './Chip'
 export * from './ChipButton'
+export * from './Drawer'
 export * from './Divider'
 export * from './Form'
 export * from './Icon'
